src/requirements.txt
git+https://github.com/klima7/signalbot
opencv-python-headless
openai
deepl
pillow
numpy
